requirements.txt
django
django-report-utils
